refactor; clean up cyber bottom bar
